DRAFT Compiler FE remove dup in chef cook
on going draft to reduce duplicate codes in circlechef and tflchef

ONE-DCO-1.0-ONE-DCO-1.0-ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
